+++ src/types/index.ts
@@ -2,9 +2,9 @@ export { ICompile } from './compile';
 export { ICompileUtilForRepeat, ICompileUtil } from './compile-utils';
 export { ComponentList, IComponent, SetState } from './component';
 export { IMiddleware, EsRouteObject, IInDiv } from './indiv';
-export { INvModule, TInjectTokenProvider, TUseClassProvider, TuseValueProvider } from './nvModule';
-export { IKeyWatcher } from './keyWatcher';
+export { INvModule, TInjectTokenProvider, TUseClassProvider, TuseValueProvider } from './nv-module';
+export { IKeyWatcher } from './key-watcher';
 export { TRouter, IRouter, GetLocation, SetLocation } from './router';
 export { IUtil } from './utils';
-export { IVnode, TAttributes, IPatchList } from './virtualDOM';
+export { IVnode, TAttributes, IPatchList } from './virtual-dom';
 export { TFnWatcher, TFnRender, IWatcher } from './watcher';
