+++ src/types/virtual-dom.ts
